+++ doc/changelog.md
@@ -19,6 +19,8 @@
 - Fixed users being able to access tasks with hashlists they would not be allowed to view.
 - Fixed users being able to access hashlists they are not allowed to see.
 - Adjusted handling to be able to deal with changed mode 22000 output.
+- Fixed pagination of hashes on cracks page.
+- Time of Zaps inserted is now saved.
 
 ## Enhancements
 
@@ -29,6 +31,7 @@
 - Allow changing the status interval for created tasks.
 - Permissions for managing access groups is separate from the permission to manage users.
 - The agent status page shows more detailed information on temperature and usage.
+- JQuery updated to v3.6.0.
 
 # v0.11.0 -> v0.12.0